Move files and convert hooks
