+++ package.json
@@ -33,6 +33,7 @@
   "dependencies": {
     "@styled-system/prop-types": "^5.1.2",
     "@testing-library/react": "^9.1.3",
+    "@types/react-transition-group": "^4.4.0",
     "chromatic": "^4.0.2",
     "cleave.js": "~1.4.10",
     "libphonenumber-js": "^1.7.13",
