+++ src/Icon/Icon.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderWithTheme } from '../../test/utils';
+import { renderWithTheme } from '../test/utils';
 import Icon from './Icon';
 
 describe('<Icon />', () => {
