+++ src/hooks/useKeyPress.ts
@@ -1,10 +1,14 @@
 import { useCallback, useEffect } from 'react';
 
-export default function useKeyPress(targetKeys, callback, keyEvent = 'keydown') {
-  const keys = [].concat(targetKeys);
-
+export const useKeyPress = (
+  targetKeys: string | string[],
+  callback: (event: KeyboardEvent) => void,
+  keyEvent = 'keydown'
+) => {
   const pressHandler = useCallback(
-    event => {
+    (event: KeyboardEvent) => {
+      const keys = Array.isArray(targetKeys) ? targetKeys : [targetKeys];
+
       if (keys.indexOf(event.key) >= 0) {
         callback(event);
       }
@@ -15,5 +19,5 @@ export default function useKeyPress(targetKeys, callback, keyEvent = 'keydown') 
   useEffect(() => {
     window.addEventListener(keyEvent, pressHandler);
     return () => window.removeEventListener(keyEvent, pressHandler);
-  }, [targetKeys, callback, keyEvent]);
-}
+  }, [pressHandler, keyEvent]);
+};
